Move _PaginatedIterable to its own submodule
